+++ NEWS.md
@@ -47,6 +47,11 @@ Standard library changes
 
 #### Profile
 
+#### Random
+
+* it's now possible to efficiently create a new `Xoshiro` RNG from an existing one via `Random.fork`, which
+  can be useful in parallel computations ([#58193]).
+
 #### REPL
 
 #### Test